data_ingestion/requirements.txt
kafka-python-ng==2.2.3
requests==2.31.0
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.1.0
python-dateutil==2.9.0
